fix(dilesa-checklist): Sprint 4C — extraer PROYECTO_DETALLE_COLUMNAS a archivo server-safe
Beto reportó HTTP 500 persistente del PDF en preview, incluso tras el
fix de `gap`. Los logs de Vercel muestran TypeError pero el mensaje
está truncado y mi try/catch no aparece — el error pasa ANTES del
handler, durante el bundle/module-load.

Causa probable: route handler importaba PROYECTO_DETALLE_COLUMNAS de
components/dilesa/proyecto-detalle.tsx que tiene 'use client'. Next.js
bundlea eagerly el árbol del client component (incluyendo
createSupabaseBrowserClient, useEffect, etc.) cuando importas
cualquier cosa de él desde un server route. Eso puede romper en
runtime con TypeErrors opacos.

Fix:
- Extraer la constante a lib/dilesa/proyecto-detalle-columnas.ts
  (puro server-safe).
- Re-exportar desde components/dilesa/proyecto-detalle.tsx por
  back-compat — los imports existentes desde components/ y app/
  siguen funcionando sin tocar.
- Actualizar route.tsx para importar del nuevo archivo lib/.
